port EmailService from Java code
This does not work entirely yet because we do not have the ability to do
callbacks from Java code. We use a helper method that will be added to
`flang_dev` for now, and will need to be migrated to this repository as
soon as we don't have any other code from `flang_dev` anymore.

Link: https://github.com/tokiwa-software/fuzion/issues/3562